Group controllers together
