+++ src/db/getLocations.js
@@ -1,28 +1,12 @@
 import axios from 'axios';
-import { createClient } from 'redis';
-
-const { CACHE_EXPIRY_SECS: expiry, CACHE_URL: url, CACHE_TLS } = process.env;
+import cache from './cache';
 
 const request = axios.create({
   baseURL: 'https://maps.googleapis.com/maps/api/geocode',
   method: 'GET',
 });
 
-const cache = createClient({
-  socket: { tls: CACHE_TLS === 'true' || CACHE_TLS === '1' },
-  url,
-});
-
-const reverseGeocode = async function reverseGeocode(id, [lng, lat]) {
-  if (!cache.isOpen) {
-    await cache.connect();
-  }
-  const key = `loc:${id}`;
-  const cached = await cache.get(key, 'EX', expiry);
-  if (cached) {
-    return JSON.parse(cached);
-  }
-
+const reverseGeocode = async function reverseGeocode([lng, lat]) {
   const { data: { results: [address] } } = await
   request(`json?key=${process.env.GOOGLE_KEY}&latlng=${lat},${lng}`);
 
@@ -40,14 +24,16 @@ const reverseGeocode = async function reverseGeocode(id, [lng, lat]) {
     location.address = `${street_number} ${route}`;
   }
 
-  await cache.setEx(key, expiry, JSON.stringify(location));
-
   return location;
+};
+
+const geocache = function geocache(id, [lng, lat]) {
+  return cache(`loc:${id}`, async () => reverseGeocode([lng, lat]));
 };
 
 export default async function getLocations(rows) {
   return Promise.all(rows.map(async (row) => {
-    row.properties.location = await reverseGeocode(
+    row.properties.location = await geocache(
       row.properties.id,
       row.geometry.coordinates,
     );
