+++ src/pages/api/event/[id].js
@@ -1,5 +1,5 @@
 import handler, { HttpException } from '../../handler';
-import getEvent from '../../../db/getEvent';
+import Event from '../../../db/event';
 
 /**
  * @param {import('http').IncomingMessage} req
@@ -9,7 +9,7 @@ const GET = async function GET(req, res) {
   const id = req.intParam('id');
   const limit = req.intParam('limit', null);
   const offset = req.intParam('offset', null);
-  const event = await getEvent(id, limit, offset);
+  const event = await Event.get(id, limit, offset);
   if (event === undefined) {
     throw new HttpException(404, 'Event not found', id);
   }
