+++ src/pages/profile/[id].js
@@ -1,8 +1,8 @@
 import Profile from '../../components/profile/Profile';
-import getProfile from '../../db/getProfile';
+import Busker from '../../db/busker';
 
 export const getStaticProps = async function getStaticProps(context) {
-  const performer = await getProfile(context.params.id);
+  const performer = await Busker.get(context.params.id);
   return performer === undefined ? { notFound: true } : { props: { performer } };
 };
